+++ app/src/main/java/com/example/pokemonotravezconfragmentos/data/api/PokemonListResponse.kt
@@ -11,8 +11,8 @@ data class PokemonListItemResponse(
 data class PokemonDetailResponse(
     val id:Int,
     val name:String,
-    val weight: Int,
-    val height: Int,
+    val weight: Double,
+    val height: Double,
     val sprites: PokemonSpritesResponse
 )
